Added a separate unit-test for zim::Search
Uncommenting of the search related functionality checks in the
ZimArchive unit test leads to the failure of the test cases
where the search functionality is being exercised (ZimArchive.openByFD
and ZimArchive.openZIMFileEmbeddedInAnotherFile). Adding a separate
unit test for zim search demonstrates that those failure aren't (fully)
caused by the implementation of the ZIM file opening via a file
descriptor.